update tms5220 core to latest upstream version

+++ BlueWizard/SpeechSynthesizer.mm
@@ -69,7 +69,7 @@
 
 -(void)fillBuffer:(NSMutableArray *)samples {
     unsigned int frames = _tms5220->m_fifo_count;
-    int buffer[frames];
+    short int buffer[frames];
     _tms5220->process(buffer, frames);
     float scale = 1.0f / (1 << 15);
     for (int i = 0; i < frames; i++) {
